app/requirements.txt
Django==3.2.8
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework==3.11.2
drf_yasg==1.20.0
pytest-django==4.4.0
